+++ api/tests/test_embeddings.py
